Rename event handler files (#574)
