+++ src/eventHandlers/WorkerInitHandler.ts
